requirements.txt
simple-salesforce==1.12.6
python-dotenv==1.1.1
requests==2.32.3

# LangGraph & LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.9
langchain-core==0.3.63

# OpenAI Realtime API
openai==1.109.1
websockets==12.0

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Utilities
pydantic==2.12.3
dataclasses-json==0.6.3

# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Database
pymongo==4.15.3

# Twilio API
twilio==9.8.4

# MCP Tools
fastmcp==2.12.5
mcp==1.16.0

# Compatible async library
anyio==4.5.0

